hw04/asr_code/requirements.txt
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy>=1.24.3